Optimize SwitchModule and don't open tab if open
This moves some functionality within the if statements for speed.
Also, it checks whether the file is already open, if so it will switch
to the file instead of opening it again. This approach is a bit crude
and might not word with relative paths compared to absolute ones.
However, this works in most use cases and doesn't break anything.