don't use %s in errorformat
Vim does not jump to the location of an error with the current
errorformat.

With %s vim tries to find the error location using what %s matches, even
adding ^ and $. When %s is present, vim ignores the line number of the
error message. The error messages from the clean compiler contain only a
snippet somehow related to the error, which may not even be in the same
line.

This patch makes errorformat ignore the snippet, by using the pattern
%*[^]], which matches everything until the next closing ]. This causes
vim to use the line number to jump to the error location.